+++ src/commands/echo_command.c
@@ -1,15 +1,10 @@
-#include "commands.h"
-#include <stdio.h>
-#include <unistd.h>
-#include <string.h>
-#include <stdlib.h>
 #include <ctype.h>
-#include <libft.h>
-#include <parser/parser.h>
+#include <commands/buffer.h>
+#include <commands/commands.h>
 #include <commands/echo_utils.h>
 #include <commands/quotes.h>
-#include <commands/buffer.h>
 #include <env/env_utils.h>
+#include <parser/parser.h>
 
 static void	handle_spaces(t_arg *arg, t_buffer *buffer)
 {
