+++ src/commands/echo_utils.h
@@ -1,8 +1,8 @@
 #ifndef ECHO_UTILS_H
 # define ECHO_UTILS_H
-# include <defines.h>
-# include <commands/commands.h>
 # include <commands/buffer.h>
+# include <commands/commands.h>
+# include <defines.h>
 
 t_bool	parse_n_flag(t_arg *input);
 void	write_to_stdout(const char *string_to_write);
