+++ src/commands/exit_command.c
@@ -1,8 +1,6 @@
-#include <commands/commands.h>
 #include <stdlib.h>
-#include <stdio.h>
+#include <commands/commands.h>
 
-/* parse argument to get return code */
 t_exit_code	exit_command(t_command command, t_output_stdout write_to_stdout)
 {
 	(void)command;
