+++ src/executor/redirection_utils.c
@@ -1,7 +1,6 @@
+#include <libft.h>
 #include <executor/run_commands.h>
 #include <parser/command_table.h>
-#include <libft.h>
-#include <stdio.h>
 
 void	handle_stdin(int in_file, t_multi_pipes *pipes, int current_process)
 {
