+++ src/parser/command_table.h
@@ -1,7 +1,7 @@
 #ifndef COMMAND_TABLE_H
 # define COMMAND_TABLE_H
-# include <defines.h>
 # include <commands/commands.h>
+# include <defines.h>
 
 t_command_code	get_command_code(const char **input, t_command *command);
 t_bool			is_between_quotes(const char *input, int reserved_char_index);
